Even more refactor to remove circular dependencies
https://coveord.atlassian.net/browse/JSUI-2725
