+++ src/ui/DynamicFacet/DynamicFacetValues/DynamicFacetValue.ts
@@ -1,11 +1,17 @@
 import * as Globalize from 'globalize';
-import { DynamicFacetValueRenderer } from './DynamicFacetValueRenderer';
 import { FacetValueState } from '../../../rest/Facet/FacetValueState';
 import { IAnalyticsDynamicFacetMeta, analyticsActionCauseList } from '../../Analytics/AnalyticsActionListMeta';
 import { l } from '../../../strings/Strings';
 import { IRangeValue, RangeType } from '../../../rest/RangeValue';
 import { FacetType } from '../../../rest/Facet/FacetRequest';
 import { IDynamicFacet } from '../DynamicFacetInterface';
+import { IFacetResponseValue } from '../../../rest/Facet/FacetResponse';
+
+export interface ValueCreator {
+  createFromResponse(facetValue: IFacetResponseValue, index: number): DynamicFacetValue;
+  createFromValue(value: string): DynamicFacetValue;
+  createFromRange?(range: IRangeValue, index: number): DynamicFacetValue;
+}
 
 export interface ValueRenderer {
   render(): HTMLElement;
@@ -37,7 +43,7 @@ export class DynamicFacetValue implements IDynamicFacetValue {
   public renderer: ValueRenderer;
   private element: HTMLElement = null;
 
-  constructor(facetValue: IDynamicFacetValue, private facet: IDynamicFacet, rendererKlass: IValueRendererKlass = DynamicFacetValueRenderer) {
+  constructor(facetValue: IDynamicFacetValue, private facet: IDynamicFacet, rendererKlass: IValueRendererKlass) {
     this.value = facetValue.value;
     this.start = facetValue.start;
     this.end = facetValue.end;
