+++ src/ui/DynamicFacet/DynamicFacetValues/DynamicFacetValueCreator.ts
@@ -1,12 +1,12 @@
 import { FacetUtils } from '../../Facet/FacetUtils';
-import { ValueCreator } from './DynamicFacetValues';
-import { DynamicFacetValue } from './DynamicFacetValue';
+import { DynamicFacetValue, ValueCreator } from './DynamicFacetValue';
 import { IFacetResponseValue } from '../../../rest/Facet/FacetResponse';
 import { FacetValueState } from '../../../rest/Facet/FacetValueState';
 import { IDynamicFacet } from '../DynamicFacetInterface';
+import { DynamicFacetValueRenderer } from './DynamicFacetValueRenderer';
 
 export class DynamicFacetValueCreator implements ValueCreator {
-  constructor(private facet: IDynamicFacet) {}
+  constructor(private facet: IDynamicFacet) { }
 
   private formatDisplayValue(value: string) {
     let returnValue = FacetUtils.tryToGetTranslatedCaption(<string>this.facet.options.field, value);
@@ -27,7 +27,8 @@ export class DynamicFacetValueCreator implements ValueCreator {
         state: facetValue.state,
         position: index + 1
       },
-      this.facet
+      this.facet,
+      DynamicFacetValueRenderer
     );
   }
 
@@ -35,6 +36,10 @@ export class DynamicFacetValueCreator implements ValueCreator {
     const position = this.facet.values.allFacetValues.length + 1;
     const state = FacetValueState.idle;
     const displayValue = this.formatDisplayValue(value);
-    return new DynamicFacetValue({ value, displayValue, state, numberOfResults: 0, position }, this.facet);
+    return new DynamicFacetValue(
+      { value, displayValue, state, numberOfResults: 0, position },
+      this.facet,
+      DynamicFacetValueRenderer
+    );
   }
 }
