+++ src/ui/DynamicFacet/DynamicFacetValues/DynamicFacetValues.ts
@@ -1,19 +1,12 @@
 import 'styling/DynamicFacet/_DynamicFacetValues';
 import { $$ } from '../../../utils/Dom';
 import { findWhere, find } from 'underscore';
-import { DynamicFacetValue } from './DynamicFacetValue';
-import { IFacetResponse, IFacetResponseValue } from '../../../rest/Facet/FacetResponse';
+import { DynamicFacetValue, ValueCreator } from './DynamicFacetValue';
+import { IFacetResponse } from '../../../rest/Facet/FacetResponse';
 import { FacetValueState } from '../../../rest/Facet/FacetValueState';
 import { l } from '../../../strings/Strings';
-import { DynamicFacetValueCreator } from './DynamicFacetValueCreator';
 import { IRangeValue } from '../../../rest/RangeValue';
 import { IDynamicFacet } from '../DynamicFacetInterface';
-
-export interface ValueCreator {
-  createFromResponse(facetValue: IFacetResponseValue, index: number): DynamicFacetValue;
-  createFromValue(value: string): DynamicFacetValue;
-  createFromRange?(range: IRangeValue, index: number): DynamicFacetValue;
-}
 
 export interface IDynamicFacetValueCreatorKlass {
   new (facet: IDynamicFacet): ValueCreator;
@@ -24,7 +17,7 @@ export class DynamicFacetValues {
   private list = $$('ul', { className: 'coveo-dynamic-facet-values' }).el;
   private valueCreator: ValueCreator;
 
-  constructor(private facet: IDynamicFacet, creatorKlass: IDynamicFacetValueCreatorKlass = DynamicFacetValueCreator) {
+  constructor(private facet: IDynamicFacet, creatorKlass: IDynamicFacetValueCreatorKlass) {
     this.resetValues();
     this.valueCreator = new creatorKlass(this.facet);
   }
