+++ unitTests/ui/DynamicFacet/DynamicFacetBreadcrumbsTest.ts
@@ -6,6 +6,7 @@ import { IDynamicFacetOptions } from '../../../src/ui/DynamicFacet/DynamicFacetI
 import { IDynamicFacetValue, DynamicFacetValue } from '../../../src/ui/DynamicFacet/DynamicFacetValues/DynamicFacetValue';
 import { FacetValueState } from '../../../src/rest/Facet/FacetValueState';
 import { analyticsActionCauseList } from '../../../src/ui/Analytics/AnalyticsActionListMeta';
+import { DynamicFacetValueRenderer } from '../../../src/ui/DynamicFacet/DynamicFacetValues/DynamicFacetValueRenderer';
 
 export function DynamicFacetBreadcrumbsTest() {
   describe('DynamicFacetBreadcrumbs', () => {
@@ -104,7 +105,7 @@ export function DynamicFacetBreadcrumbsTest() {
         $$(valueElements()[0]).trigger('click');
         expect(facet.logAnalyticsEvent).toHaveBeenCalledWith(
           analyticsActionCauseList.breadcrumbDynamicFacet,
-          new DynamicFacetValue(mockFacetValues[0], facet).analyticsMeta
+          new DynamicFacetValue(mockFacetValues[0], facet, DynamicFacetValueRenderer).analyticsMeta
         );
       });
     });
